introduzindo o Sistema de Mesas

+++ src/main/java/Model/repositories/Estoque.java
@@ -8,7 +8,7 @@ import Model.exceptions.ProdutoNaoExisteException;
 import java.util.ArrayList;
 
 public interface Estoque {
-   Products getProductById(int id);
+   Products getProductById(double id);
    boolean produtoExiste(Products product);
    void addEstoque(Products product) throws ProdutoJaExisteException;
    void removeEstoque(Products product, int id) throws ProdutoNaoExisteException;
